requirements.txt
SQLAlchemy-Utils==0.36.8
Flask-Limiter==3.1.0
PyMySQL==1.0.2
Flask==2.2.0
Werkzeug==2.2.2
requests==2.28.2
oauthlib==3.2.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.28.1
flask-marshmallow==0.14.0
Jinja2==3.1.2
itsdangerous==2.1.2
Flask-API==2.0
Flask-WTF==1.1.0
simplejson==3.18.1
Flask-OAuth==0.12
Flask-Admin==1.6.0
python-dotenv==0.21.0
wheel==0.38.4
uwsgi==2.0.21
cryptography==39.0.0
bootstrap-flask==2.2.0
flask_cors==3.0.10
